Move Teams logic to its own module
